feat(sdk-metrics-base): remove per-meter config on `MeterProvider.getMeter`
Makes it compliant with api definition.
